fix tests data files searching
Now we can run tests from anywere and all source files needed for that
will be found succesfully. Drawback of this approach is that data files
needed only for tests, and unnecessary for the main executable will be
installed with main executable. In cabal file there is only global
"data-files" declaration and we can not apply that (or similar)
declaration specially for tests only.